Done update class UI, alert UI

+++ src/Layouts/components/CourseSideBarTeacher/CourseTeacher/CourseTeacher.js
@@ -4,6 +4,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
     faChevronDown,
     faChevronUp,
+    faSpinner,
     faSquarePlus,
     faTriangleExclamation,
 } from '@fortawesome/free-solid-svg-icons';
@@ -15,9 +16,10 @@ import { teacher } from '~/redux/reducer/TeacherSlice';
 import CreateCourse from '../CreateCourse';
 import LessonTeacher from '../../CourseSidebar/LessonTeacher/LessonTeacher';
 import { activeLesson } from '~/redux/reducer/ActiveLesson';
-import { Modal } from 'antd';
+import { Button, Modal } from 'antd';
 import { useParams } from 'react-router-dom';
 import { deleteCourseUpdate } from '~/requestApi/requestUpdateClass';
+import { toastify } from '~/utils/toast';
 
 const cx = classNames.bind(styles);
 
@@ -29,20 +31,28 @@ function CourseTeacher({ data, index }) {
     const [showLesson, setShowLesson] = useState(false);
     const [showUpdateCourse, setShowUpdateCourse] = useState(false);
     const [moreEdit, setMoreEdit] = useState(false);
+    const [deleteCourseLoading, setDeleteCourseLoading] = useState(false);
 
     const lessonData = lessons.filter((item, index) => item.course_id === data.id);
 
     const handleDeleteCourse = () => {
         if (Object.keys(slug).length > 0) {
             if (Number(data.id)) {
-                deleteCourseUpdate(data.id)
-                    .then((res) => {
-                        dispatch(teacher.actions.deleteCourse(data.id));
-                        showAlertModal(false);
-                    })
-                    .catch((error) => {
-                        console.log(error);
-                    });
+                setDeleteCourseLoading(true);
+                if (!deleteCourseLoading) {
+                    deleteCourseUpdate(data.id)
+                        .then((res) => {
+                            dispatch(teacher.actions.deleteCourse(data.id));
+                            setShowAlertModal(false);
+                            setDeleteCourseLoading(false);
+                        })
+                        .catch((error) => {
+                            console.log(error);
+
+                            toastify('Delete course faild', 'error', 2000, 'top-right');
+                            setDeleteCourseLoading(false);
+                        });
+                }
             } else {
                 dispatch(teacher.actions.deleteCourse(data.id));
                 showAlertModal(false);
@@ -107,8 +117,19 @@ function CourseTeacher({ data, index }) {
             <Modal
                 title={<div className={cx('custom-modal-title')}>Warning</div>}
                 open={showAlertModal}
-                onOk={handleDeleteCourse}
                 onCancel={() => setShowAlertModal(false)}
+                footer={[
+                    <Button key="back" onClick={() => setShowAlertModal(false)}>
+                        Cancle
+                    </Button>,
+                    <Button key="submit" type="primary" onClick={handleDeleteCourse}>
+                        {deleteCourseLoading ? (
+                            <FontAwesomeIcon icon={faSpinner} className="fa-solid fa-spinner fa-spin-pulse" />
+                        ) : (
+                            'Delete'
+                        )}
+                    </Button>,
+                ]}
             >
                 <div className={cx('alert-wrap')}>
                     <span className={cx('alert-icon')}>
@@ -118,7 +139,7 @@ function CourseTeacher({ data, index }) {
                             className="fa-xl"
                         />
                     </span>
-                    <span className={cx('alert-desc')}>Du lieu cua ban se mat het!</span>
+                    <span className={cx('alert-desc')}>Du liệu của bạn sẽ mất hết!</span>
                 </div>
             </Modal>
         </>
